Refactoring code

+++ app/src/main/java/com/example/composition/domain/entity/GameSettings.kt
@@ -7,6 +7,6 @@ import kotlinx.parcelize.Parcelize
 data class GameSettings(
     val maxSumValue: Int,
     val minCountOfRightAnswers: Int,
-    val minPercentOfRightAnswers: Int,
+    val minPercentOfRightAnswers: Double,
     val gameTimeInSeconds: Int,
 ) : Parcelable
